requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==24.3.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
cffi==1.17.1
charset-normalizer==2.0.12
cryptography==44.0.0
cytoolz==0.11.2
eip712-structs==1.1.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
exceptiongroup==1.2.2
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.3.0
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==2.2.2
packaging==21.3
pandas==2.2.3
parsimonious==0.8.1
pluggy==1.0.0
prometheus-client==0.14.1
protobuf==3.19.4
py==1.11.0
py_clob_client==0.17.1
py_order_utils==0.3.0
pycparser==2.22
pycryptodome==3.14.1
pyparsing==3.0.7
pyrsistent==0.18.1
pysha3==1.0.2
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0
requests==2.27.1
rlp==2.0.1
six==1.16.0
tomli==2.0.1
toolz==0.11.2
tzdata==2025.1
urllib3==1.26.8
varint==1.0.2
web3==5.28.0
websockets==9.1
yarl==1.7.2
